Refactored as per PR comments
